CR - StreamMetadatas
