+++ atlasdb-client/src/main/java/com/palantir/atlasdb/stream/StreamMetadatas.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2018 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the BSD-3 License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.atlasdb.stream;
+
+import com.palantir.atlasdb.protos.generated.StreamPersistence;
+
+public final class StreamMetadatas {
+    private StreamMetadatas() {
+        // utility class
+    }
+
+    public static long getNumberOfBlocksFromMetadata(StreamPersistence.StreamMetadata metadata) {
+        return (metadata.getLength() + GenericStreamStore.BLOCK_SIZE_IN_BYTES - 1)
+                / GenericStreamStore.BLOCK_SIZE_IN_BYTES;
+    }
+}
