+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/sweep/external/SchemalessStreamStoreDeleterTest.java
@@ -22,21 +22,15 @@ import java.util.Arrays;
 
 import org.junit.Test;
 
-import com.google.protobuf.ByteString;
 import com.palantir.atlasdb.keyvalue.api.Namespace;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.protos.generated.StreamPersistence;
 import com.palantir.atlasdb.schema.cleanup.ImmutableStreamStoreCleanupMetadata;
 import com.palantir.atlasdb.schema.stream.StreamTableType;
-import com.palantir.atlasdb.stream.GenericStreamStore;
 import com.palantir.atlasdb.table.description.ValueType;
 
 public class SchemalessStreamStoreDeleterTest {
-    private static final ByteString BYTE_STRING = ByteString.copyFrom(new byte[] { 1, 2, 3 });
     private static final Namespace NAMESPACE = Namespace.DEFAULT_NAMESPACE;
     private static final String SHORT_NAME = "s";
-
-    private static final int BLOCK_SIZE = GenericStreamStore.BLOCK_SIZE_IN_BYTES;
 
     private final SchemalessStreamStoreDeleter deleter = new SchemalessStreamStoreDeleter(
             NAMESPACE,
@@ -51,40 +45,5 @@ public class SchemalessStreamStoreDeleterTest {
         Arrays.stream(StreamTableType.values()).forEach(type ->
                 assertThat(deleter.getTableReference(type))
                         .isEqualTo(TableReference.create(NAMESPACE, type.getTableName(SHORT_NAME))));
-    }
-
-    @Test
-    public void getNumberOfBlocksFromMetadata_Normal() {
-        assertThat(getNumberOfBlocksFromDeleter(3 * BLOCK_SIZE + BLOCK_SIZE / 2)).isEqualTo(4);
-        assertThat(getNumberOfBlocksFromDeleter(1)).isEqualTo(1);
-    }
-
-    @Test
-    public void getNumberOfBlocksFromMetadata_ExactMatch() {
-        assertThat(getNumberOfBlocksFromDeleter(10 * BLOCK_SIZE)).isEqualTo(10);
-        assertThat(getNumberOfBlocksFromDeleter(744 * BLOCK_SIZE)).isEqualTo(744);
-    }
-
-    @Test
-    public void getNumberOfBlocksFromMetadata_OffByOnes() {
-        assertThat(getNumberOfBlocksFromDeleter(10 * BLOCK_SIZE - 1)).isEqualTo(10);
-        assertThat(getNumberOfBlocksFromDeleter(10 * BLOCK_SIZE + 1)).isEqualTo(11);
-    }
-
-    @Test
-    public void getNumberOfBlocksFromMetadata_Zero() {
-        assertThat(getNumberOfBlocksFromDeleter(0)).isEqualTo(0);
-    }
-
-    private long getNumberOfBlocksFromDeleter(int length) {
-        return deleter.getNumberOfBlocksFromMetadata(createStreamMetadataWithLength(length));
-    }
-
-    private static StreamPersistence.StreamMetadata createStreamMetadataWithLength(int length) {
-        return StreamPersistence.StreamMetadata.newBuilder()
-                .setHash(BYTE_STRING)
-                .setLength(length)
-                .setStatus(StreamPersistence.Status.STORED)
-                .build();
     }
 }
